refactor: split data.rs into own dir
