+++ src/data/mod.rs
@@ -1,0 +1,4 @@
+pub mod data;
+pub mod genotypes;
+pub mod phen_stats;
+pub mod phenotypes;
